// src/app/shared/enums/routes.ts
export enum RouteEnum {
  AdminDashboard = '/admin/dashboard',
  AdminProfiles = '/admin/profiles',
  AdminNudgeStudents = '/admin/nudgeStudents',
  AdminFundAllocations = '/admin/fundAllocations',
  AdminDeclinedApplications = '/admin/declinedApplications',
  HodDashboardBursaries = '/dashboard',
  HodDashboardApplication = '/dashboard',
  DeanFunds = 'dashboard/funds',
  Reports = '/admin/reports',
  OnboardStudents = '/admin/onboardStudents',
  AdminActiveBursaries = '/admin/activeBursaries',
  AdminApplications = '/admin/applications',
  AdminCreateEvent = '/admin/createEvent',
  AdminEventList = '/admin/events'
}